QuadraticConverter.jl Documentation

QuadraticConverter.to_expanded — Method

Converts the factored function into expanded form

QuadraticConverter.to_expanded — Method

Converts the vertex function into expanded form

QuadraticConverter.to_factored — Method

Converts the expanded function into factored form

QuadraticConverter.to_factored — Method

Converts the vertex function into factored form

QuadraticConverter.to_vertex — Method

Converts the expanded function into vertex form

QuadraticConverter.to_vertex — Method

Converts the factored function into vertex form

QuadraticConverter.value — Method

Returns the result of the function applied to the second parameter

QuadraticConverter.value — Method

Returns the result of the function applied to the second parameter

QuadraticConverter.value — Method

Returns the result of the function applied to the second parameter

QuadraticConverter.expanded — Type

a, b, c

QuadraticConverter.vertex — Type

a, p, q

QuadraticConverter.factored — Type

a, x1, x2